Rename directories and update chat screen
